+++ RELEASES.md
@@ -3,7 +3,10 @@
 ## 0.2.1 (unreleased)
 
 ### Fixes
+
 - `taffy::compute_layout` has been made public allowing Taffy to be used with custom storage (#263)
+
+- In case of conflicts, `min_size` now overrides `max_size` which overrides `size` (#261)
 
 ## 0.2.0